+++ spec/configuration/sidekiq_scheduled_jobs_spec.rb
@@ -27,6 +27,7 @@ RSpec.describe "Sidekiq configuration" do
       SetOrganisationSlugsJob
       SetOrganisationSlugsOfBatchJob
       ImportFromVacancySourcesJob
+      ImportFromVacancySourceJob
     ]
   end
